Don't show notif if we rebooted in last 24hr.
Currently, the boot notif appears at a set time each day, if
aconfig flags are staged that day. With this change, it's the
same, except that when it's time to show a notification, we
check if it's been at least 24 hours since the last reboot.
If it has, we show the notification, otherwise we postpone to
the next day.

Test: tested manually on device
Bug: 306199946
Change-Id: I7f0ff9f70f1f49fe4642cfa2f63f5bfdecbca12e