Move DeviceConfigServiceResources to impl_only_libs
DeviceConfigServiceResources is only required to compile the
implementation of this java_sdk_library. `service-configinfrastructure`
contributes to the system_server_classpath; which is used to create an
api_fingeerprint.txt file. To prevent circular deps, move this out of
libs(used to compile impl+stubs) to impl_only_libs (used to compile
impl)

Bug: 312465394
Test: m service-configinfrastructure
Test: TARGET_RELEASE=trunk TARGET_BUILD_VARIANT=userdebug UNBUNDLED_BUILD_SDKS_FROM_SOURCE=true UNBUNDLED_BUILD_TARGET_SDK_WITH_API_FINGERPRINT=true TARGET_BUILD_APPS=com.google.android.sdkext vendor/google/build/mainline_modules_x86_64.sh -j8
Change-Id: I268f4c8ceb3deb8cc650bf3074e71cfd4026ff05