Update aflags to work with mainline beta flags
aflags currently only interact with new storage. For it to work with
mainline beta flags which are stored in both old and new storage. We
need to update aflags to also call into device config when it is a
mainline beta flag.

This CL does two things primarily:

1, capture which flags are mainline beta flags. Aflags command loads all
aconfig_flags.proto in each command call. The storage backend field in
the loaded parsed flag proto can be used to determine if it also needs
to interact with device config.

2, added device_config_source module which implements FlagSource trait.
It calls into device_config system command.

There will be another cl to connect device config commands in the main.

Bug: b/406510373
Test: added unit tests.
Flag: EXEMPT flagging infra
Change-Id: Ie37d2403eb2193589d39042ea05c074f1992dc9f